+++ server/test/unit/vitalsTest/checkGatewayHeartbeatTest.js
@@ -7,7 +7,7 @@ const rewire = require('rewire')
 const { DateTime } = require('luxon')
 
 // In-house dependencies
-const { helpers } = require('brave-alert-lib')
+const { factories, helpers } = require('brave-alert-lib')
 const db = require('../../../db/db')
 const { gatewayFactory, gatewaysVitalFactory } = require('../../testingHelpers')
 const aws = require('../../../aws')
@@ -57,9 +57,9 @@ describe('vitals.js unit tests: checkGatewayHeartbeat', () => {
     sandbox.restore()
   })
 
-  describe('when last seen gateway is less than the threshold and no alerts have been sent ', () => {
+  describe('when gateway that is sending vitals last seen is less than the threshold and no alerts have been sent ', () => {
     beforeEach(async () => {
-      this.gateway = gatewayFactory({ sentVitalsAlertAt: null })
+      this.gateway = gatewayFactory({ sentVitalsAlertAt: null, isSendingVitals: true, client: factories.clientFactory({ isSendingVitals: true }) })
       sandbox.stub(db, 'getGateways').returns([this.gateway])
       this.gatewaysVital = gatewaysVitalFactory({
         lastSeenAt: currentDBDate,
@@ -84,9 +84,13 @@ describe('vitals.js unit tests: checkGatewayHeartbeat', () => {
     })
   })
 
-  describe('when last seen gateway is less than the threshold and an alert was sent ', () => {
+  describe('when gateway that is sending vitals last seen is less than the threshold and an alert was sent ', () => {
     beforeEach(async () => {
-      this.gateway = gatewayFactory({ sentVitalsAlertAt: currentDBDate })
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: currentDBDate,
+        isSendingVitals: true,
+        client: factories.clientFactory({ isSendingVitals: true }),
+      })
       sandbox.stub(db, 'getGateways').returns([this.gateway])
       this.gatewaysVital = gatewaysVitalFactory({
         lastSeenAt: currentDBDate,
@@ -115,9 +119,102 @@ describe('vitals.js unit tests: checkGatewayHeartbeat', () => {
     })
   })
 
-  describe('when last seen gateway is more than the threshold and no alerts have been sent', () => {
+  describe('when gateway that is not sending vitals last seen is less than the threshold and an alert was sent ', () => {
     beforeEach(async () => {
-      this.gateway = gatewayFactory({ sentVitalsAlertAt: null })
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: currentDBDate,
+        isSendingVitals: false,
+        client: factories.clientFactory({ isSendingVitals: true }),
+      })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: currentDBDate,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the reconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not send a reconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to null", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
+  describe('when gateway that is sending vitals but whose client is not sending vitals last seen is less than the threshold and an alert was sent ', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: currentDBDate,
+        isSendingVitals: true,
+        client: factories.clientFactory({ isSendingVitals: false }),
+      })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: currentDBDate,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the reconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not send a reconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to null", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
+  describe('when gateway that is not sending vitals and whose client is also not sending vitals last seen is less than the threshold and an alert was sent ', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: currentDBDate,
+        isSendingVitals: false,
+        client: factories.clientFactory({ isSendingVitals: false }),
+      })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: currentDBDate,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the reconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not send a reconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to null", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
+  describe('when gateway that is sending vitals last seen is more than the threshold and no alerts have been sent', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({ sentVitalsAlertAt: null, isSendingVitals: true, client: factories.clientFactory({ isSendingVitals: true }) })
       sandbox.stub(db, 'getGateways').returns([this.gateway])
       this.gatewaysVital = gatewaysVitalFactory({
         lastSeenAt: exceededTimestamp,
@@ -146,9 +243,94 @@ describe('vitals.js unit tests: checkGatewayHeartbeat', () => {
     })
   })
 
+  describe('when gateway that is not sending vitals last seen is more than the threshold and no alerts have been sent', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({ sentVitalsAlertAt: null, isSendingVitals: false, client: factories.clientFactory({ isSendingVitals: true }) })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: exceededTimestamp,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the disconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not.send a disconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to now", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
+  describe('when gateway that is sending vitals but whose client is not sending vitals last seen is more than the threshold and no alerts have been sent', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({ sentVitalsAlertAt: null, isSendingVitals: true, client: factories.clientFactory({ isSendingVitals: false }) })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: exceededTimestamp,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the disconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not.send a disconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to now", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
+  describe('when gateway that is not sending vitals and whose client is also not sending vitals last seen is more than the threshold and no alerts have been sent', () => {
+    beforeEach(async () => {
+      this.gateway = gatewayFactory({ sentVitalsAlertAt: null, isSendingVitals: false, client: factories.clientFactory({ isSendingVitals: false }) })
+      sandbox.stub(db, 'getGateways').returns([this.gateway])
+      this.gatewaysVital = gatewaysVitalFactory({
+        lastSeenAt: exceededTimestamp,
+        gateway: this.gateway,
+      })
+      sandbox.stub(db, 'getRecentGatewaysVitalWithGatewayId').returns(this.gatewaysVital)
+    })
+
+    it('should not send the disconnection message to Sentry', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(helpers.logSentry).not.to.be.called
+    })
+
+    it('should not.send a disconnection notifications', async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(this.sendNotificationStub).not.to.be.called
+    })
+
+    it("should not update the gateway's sentVitalsAlertAt the database to now", async () => {
+      await vitals.checkGatewayHeartbeat()
+      expect(db.updateGatewaySentVitalsAlerts).not.to.be.called
+    })
+  })
+
   describe('when last seen gateway is more than the threshold and the last alert was sent less than the subsequent alert threshold', () => {
     beforeEach(async () => {
-      this.gateway = gatewayFactory({ sentVitalsAlertAt: currentDBDate })
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: currentDBDate,
+        isSendingVitals: true,
+        client: factories.clientFactory({ isSendingVitals: true }),
+      })
       sandbox.stub(db, 'getGateways').returns([this.gateway])
       this.gatewaysVital = gatewaysVitalFactory({
         lastSeenAt: exceededTimestamp,
@@ -175,7 +357,11 @@ describe('vitals.js unit tests: checkGatewayHeartbeat', () => {
 
   describe('when last seen gateway is more than the threshold and the last alert was sent more than the subsequent alert threshold', () => {
     beforeEach(async () => {
-      this.gateway = gatewayFactory({ sentVitalsAlertAt: exceededReminderTimestamp })
+      this.gateway = gatewayFactory({
+        sentVitalsAlertAt: exceededReminderTimestamp,
+        isSendingVitals: true,
+        client: factories.clientFactory({ isSendingVitals: true }),
+      })
       sandbox.stub(db, 'getGateways').returns([this.gateway])
       this.gatewaysVital = gatewaysVitalFactory({
         lastSeenAt: exceededTimestamp,